requirements.txt
Flask
python-dotenv
Authlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
requests